tbi(deploy-tooling): move replaceEnvVariables to ui5-config
