+++ .changeset/many-beers-knock.md
@@ -1,0 +1,6 @@
+---
+'@sap-ux/deploy-tooling': patch
+'@sap-ux/ui5-config': patch
+---
+
+move replaceEnvVariables to ui5-config and export
