+++ packages/deploy-tooling/src/ui5/index.ts
@@ -2,9 +2,10 @@ import type { TaskParameters } from '@ui5/builder';
 import { LogLevel, ToolsLogger, UI5ToolingTransport } from '@sap-ux/logger';
 import type { AbapDeployConfig } from '../types';
 import { NAME } from '../types';
-import { deploy, replaceEnvVariables, validateConfig } from '../base';
+import { deploy, validateConfig } from '../base';
 import { createUi5Archive } from './archive';
 import { config as loadEnvConfig } from 'dotenv';
+import { replaceEnvVariables } from '@sap-ux/ui5-config';
 
 /**
  * Custom task to upload the build result to the UI5 ABAP Repository.
