+++ packages/ui5-config/src/utils.ts
@@ -73,3 +73,20 @@ export function getTypesPackage(ui5Version?: string) {
         return UI5_DEFAULT.TYPES_PACKAGE_NAME;
     }
 }
+
+/**
+ * Replace environment variable references of pattern `env:VAR_NAME` with the value of the corresponding environment variable.
+ *
+ * @param obj - any object structure
+ */
+export function replaceEnvVariables(obj: object): void {
+    for (const key in obj) {
+        const value = (obj as Record<string, unknown>)[key];
+        if (typeof value === 'object') {
+            replaceEnvVariables(value as object);
+        } else if (typeof value === 'string' && value.startsWith('env:')) {
+            const varName = value.split('env:')[1];
+            (obj as Record<string, unknown>)[key] = process.env[varName];
+        }
+    }
+}
