+++ packages/ui5-config/test/utils.test.ts
@@ -1,5 +1,5 @@
 import { UI5_DEFAULT } from '../src/defaults';
-import { getEsmTypesVersion, getTypesPackage, getTypesVersion, mergeObjects } from '../src/utils';
+import { getEsmTypesVersion, getTypesPackage, getTypesVersion, mergeObjects, replaceEnvVariables } from '../src/utils';
 
 describe('mergeObjects', () => {
     const base = {
@@ -124,5 +124,27 @@ describe('getEsmTypesVersion, getTypesVersion, getTypesPackage', () => {
     // Tests validation of getting the correct types package name
     test.each(getTypesPackageTestData)('Types package: (%p, %p)', (input, expected) => {
         expect(getTypesPackage(input)).toEqual(expected);
+    });
+});
+describe('replaceEnvVariables', () => {
+    const envVal = '~testvalue';
+    const envRef = 'env:TEST_VAR';
+
+    process.env.TEST_VAR = envVal;
+    test('top level', () => {
+        const config = { hello: envRef };
+        replaceEnvVariables(config);
+        expect(config.hello).toBe(envVal);
+    });
+    test('in array', () => {
+        const config = ['hello', envRef];
+        replaceEnvVariables(config);
+        expect(config[1]).toBe(envVal);
+    });
+    test('nested', () => {
+        const config = { hello: { world: envRef }, world: envRef };
+        replaceEnvVariables(config);
+        expect(config.hello.world).toBe(envVal);
+        expect(config.world).toBe(envVal);
     });
 });
